Add User model and users table
