+++ .env.example
@@ -1,0 +1,5 @@
+# Lito environment variables (copy this file to .env)
+ENV=dev
+
+# Database
+DATABASE_URL=postgresql://lito:lito@localhost:5432/lito
