--- apps/api/alembic/versions/3a8f791ef1f2_create_ping_table.py
@@ -1,32 +1,0 @@
-"""Create ping table
-
-Revision ID: 3a8f791ef1f2
-Revises: ee1fb890d451
-Create Date: 2026-01-20 02:56:03.334479
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3a8f791ef1f2'
-down_revision: Union[str, Sequence[str], None] = 'ee1fb890d451'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
