+++ apps/api/alembic/versions/ee1fb890d451_create_ping_table.py
@@ -13,7 +13,7 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision: str = 'ee1fb890d451'
-down_revision: Union[str, Sequence[str], None] = 'ac1441fc07a6'
+down_revision: Union[str, Sequence[str], None] = None
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
