+++ compose.yml
@@ -10,6 +10,8 @@ services:
       - db
     volumes:
       - ./apps/api:/app
+    env_file:
+      - .env
 
   db:
     image: postgres:16
